refactor: remove useless async_runtime module
